Dracula Theme
A theme created based on the official Dracula Theme palette found https://draculatheme.com/contribute/. I do not own Dracula, but thought it would be nice to share this with others.

Added a mod.js file as well as a folder for the theme files

The theme works outside of the developer build.

Misunderstood what you meant by license field.

Just corrected it.