Fix margins for every emoji set (#585)
Custom emojis, when displayed as icons in every page and in the left sidebar of Notion, are a bit misaligned. This is not the case for the emojis in the page content. Just specific places when they act like icons. They are a few pixels off but it triggers an OCD in me 🤣

This fixes it and it looks awesome. Precisely aligned as the default emoji set.